Alpha port changes.  These are not in final form.  They're hacky in the sense that I just replaced "int" with "long" and "unsigned" with "unsigned long" in places where pointers were extracted from integers and vice versa.  This should be replaced with a #define ADDRESS_TYPE unsigned long or somesuch. --gepr
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@85 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/activity/Schedule.h
@@ -110,7 +110,7 @@ extern void _activity_insertAction( Schedule_c *, timeval_t, CAction * );
 @public
   Activity_c  *swarmActivity;    // controlling swarm activity, if any
   id          mergeAction;       // merge action of swarm activity, if any
-  int         activationNumber;  // sequential id of activation within swarm
+  long         activationNumber;  // sequential id of activation within swarm
 }
 /*** methods in ScheduleActivity_c (inserted from .m file) ***/
 - (timeval_t) getCurrentTime;
